issue #169: updating Vault of Satoshi reported currencies to use authenticated /info/currency endpoint
git-svn-id: https://openclerk.googlecode.com/svn/trunk@1065 b12a1112-ca08-e5dd-f792-535eff3abf3b

+++ jobs/reported_currencies/vaultofsatoshi.php
@@ -2,32 +2,30 @@
 
 /**
  * Vault of Satoshi reported currencies job (#121).
- * Vault of Satoshi does not actually provide an API for supported coins, so we have to do screen scraping.
+ * Vault of Satoshi provides /info/currency which requires an authenticated endpoint.
  */
 
-$bitnz = array();
-require(__DIR__ . '/../../vendor/soundasleep/html5lib-php/library/HTML5/Parser.php');
-$html = crypto_get_contents(crypto_wrap_url("https://www.vaultofsatoshi.com/charts"));
-$dom = HTML5_Parser::parse($html);
+require(__DIR__ . "/../_vaultofsatoshi.php");
 
-// now load as XML
-$xml = new SimpleXMLElement($dom->saveXML());
+$data = vaultofsatoshi_query(get_site_config('vaultofsatoshi_info_currency_api_key'), get_site_config('vaultofsatoshi_info_currency_api_secret'), "/info/currency");
 
-// find all currencies
-$currencies = array();
-$x = $xml->xpath('//select[contains(@id,"coinPaymentCurrencies") or contains(@id,"coinOrderCurrencies")]/option');
-crypto_log("Found currencies " . print_r($x, true));
-foreach ($x as $option) {
-	$currencies[] = strtolower((string) $option["val"]);
+if (isset($data['message']) && $data['message']) {
+	throw new ExternalAPIException(htmlspecialchars($balance['message']));
 }
-$currencies = array_unique($currencies);
-crypto_log("Found currencies " . implode(", ", $currencies));
 
 // update the database
 $q = db()->prepare("DELETE FROM reported_currencies WHERE exchange=?");
 $q->execute(array($exchange['name']));
 
-foreach ($currencies as $cur) {
+foreach ($data['data'] as $row) {
+	$currency = strtolower($row['code']);
+	crypto_log("Found currency $currency");
+
+	if (!$row['tradeable']) {
+		crypto_log("Currency is not tradeable: bailing");
+		continue;
+	}
+
 	$q = db()->prepare("INSERT INTO reported_currencies SET exchange=?, currency=?");
-	$q->execute(array($exchange['name'], $cur));
+	$q->execute(array($exchange['name'], $currency));
 }
